pyproject.toml
[tool.poetry]
name = "prostoemenu-dev"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "prostoemenu_dev"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.4"
djangorestframework = "^3.14.0"
drf-spectacular = "^0.26.4"
black = "^23.7.0"
ruff = "^0.0.285"
python-dotenv = "^1.0.0"
psycopg2-binary = "^2.9.7"
django-filter = "^23.2"
django-cors-headers = "^4.2.0"
django-debug-toolbar = "^4.2.0"
loguru = "^0.7.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
